feat: Add optimization feature for composite cursors
Current implementation of pagination on composite cursors is rather
inneficient on large tables even in the presence of index. Changes in
this commit add an optimization option that causes paginator to emit a
bit different WHERE condition that query optimizators have an easier
time optimizing.

We kept changes backward-compatible. We need to opt-in to get the new
feature enabled.